Move model files
