+++ tabarena/tabarena/models/ebm/info.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tabarena.benchmark.models.ag.ebm.ebm_model import ExplainableBoostingMachineModel
+from tabarena.models.ebm.model import ExplainableBoostingMachineModel
 from tabarena.models._model_info import ModelInfo
 from tabarena.models.ebm.hpo import gen_ebm
 from tabarena.nips2025_utils.artifacts.method_metadata import MethodMetadata
