+++ tabarena/tabarena/models/perpetual_booster/hpo.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from autogluon.common.space import Categorical
 
-from tabarena.benchmark.models.ag.perpetual_booster.perpetual_booster_model import (
+from tabarena.models.perpetual_booster.model import (
     PerpetualBoosterModel,
 )
 from tabarena.utils.config_utils import ConfigGenerator
