+++ tabarena/tabarena/models/tabpfnv2_5/hpo.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from autogluon.common.space import Categorical
 
-from tabarena.benchmark.models.ag.tabpfnv2_5.tabpfnv2_5_model import (
+from tabarena.models.tabpfnv2_5.model import (
     RealTabPFNv25Model,
     TabPFNv26Model,
 )
