+++ tabarena/tabarena/models/tabstar/info.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from tabarena.benchmark.models.ag.tabstar.tabstar_model import TabSTARModel
+from tabarena.models.tabstar.model import TabSTARModel
 from tabarena.models._model_info import ModelInfo
 from tabarena.models.tabstar.hpo import gen_tabstar
 from tabarena.nips2025_utils.artifacts.method_metadata import MethodMetadata
